Tabs refactoring
