+++ src/pages/main-page/main-page.tsx
@@ -1,7 +1,7 @@
 import PromoFilmCard from '../../components/promo-film-card/promo-film-card';
 import MainPageCatalog from '../../components/main-page-catalog/main-page-catalog';
 import {Film} from '../../types/film-type';
-import {PromoFilm} from '../../types/promo-film-type';
+import {PromoFilm} from '../../types/film-type';
 import {ReactElement} from 'react';
 
 export const Genres = ['All genres-list', 'Comedies', 'Crime', 'Documentary', 'Dramas', 'Horror',
