+++ src/types/film-type.ts
@@ -3,3 +3,11 @@ export type Film = {
   title: string;
   src: string;
 };
+
+export type PromoFilm = {
+  title: string;
+  genre: string;
+  year: string;
+  imgSrc: string;
+  bgImgSrc: string;
+};
