--- src/types/promo-film-type.ts
@@ -1,7 +1,0 @@
-export type PromoFilm = {
-    title: string;
-    genre: string;
-    year: string;
-    imgSrc: string;
-    bgImgSrc: string;
-};
